fix: resolve CI compilation and linting errors
- Fix Java type conversion error (MIN_SCORE_THRESHOLD double to int)
- Fix frontend linting errors:
  - Remove unused session variables
  - Fix useEffect dependencies with useCallback
  - Replace img tags with Next.js Image component
  - Remove unused QuestionnaireAnswer interface
  - Fix any type with proper Question interface
- Update CI workflows to remove ai-service and Kafka references
- Remove duplicate fetchMessages function in chat page

+++ backend/src/main/java/com/devign/chattr/service/IntroductionsService.java
@@ -23,7 +23,7 @@ import java.util.stream.Collectors;
 public class IntroductionsService {
 
     private static final int MAX_DAILY_INTRODUCTIONS = 3;
-    private static final int MIN_SCORE_THRESHOLD = 0.5; // Minimum compatibility score
+    private static final double MIN_SCORE_THRESHOLD = 0.5; // Minimum compatibility score
     
     @Autowired
     private UserRepository userRepository;
